requirements.txt
python-twitter
facebook-sdk
watson-developer-cloud
pyyaml
python-dateutil
requests